Rework mapper, dto, service, controller v0.1

+++ Dados/src/main/java/cat/itacademy/barcelonactiva/Sulzbach/Lucas/s05/t02/n01/Dados/model/dto/PlayerDTO.java
@@ -9,9 +9,9 @@ public class PlayerDTO {
 
     private Integer userId;
     private String name;
-    private Float percentage;
+    private Double percentage;
 
-    public PlayerDTO(Integer userId, String name, Float percentage) {
+    public PlayerDTO(Integer userId, String name, Double percentage) {
         this.userId = userId;
         this.name = name;
         this.percentage = percentage;
